Refactored everything, removed unused import from setup, updated readme to include new way
